+++ src/components/AdminPage/AdminCategory.jsx
@@ -1,13 +1,10 @@
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
-import {
-  Box,
-  PlusBtn,
-  Btn,
-  DeleteBtn,
-  PopupBox,
-} from "./adminItems/AdminButtonCSS";
+import { PlusBtn, Btn, DeleteBtn } from "./adminItems/AdminButtonCSS";
+import { Box } from "./adminItems/AdminContainerCSS";
 import AllCategory from "./DummyData/AllCategory.json";
+import AddCategoryModal from "./adminItems/AddCategory";
+import EditCategoryModal from "./adminItems/EditCategory";
 
 export default function AdminCategory() {
   const allCategotyData = AllCategory; //백으로부터 get하기
@@ -78,65 +75,3 @@ export default function AdminCategory() {
     </>
   );
 }
-
-const AddCategoryModal = ({ onClose }) => {
-  const [categoryName, setCategoryName] = useState("");
-
-  const handleCategoryNameChange = (e) => {
-    setCategoryName(e.target.value);
-  };
-
-  const handleAddCategory = () => {
-    console.log(`${categoryName}카테고리가 추가 됨`);
-    //새로운 카테고리를 서버로 save
-    onClose();
-  };
-
-  return (
-    <PopupBox>
-      <h2>Add Category</h2>
-      <input
-        type="text"
-        value={categoryName}
-        onChange={handleCategoryNameChange}
-      />
-      <button onClick={handleAddCategory}>Add</button>
-      <button onClick={onClose}>Cancel</button>
-    </PopupBox>
-  );
-};
-
-const EditCategoryModal = ({
-  selectedCategoryId,
-  selectedCategoryName,
-  onClose,
-}) => {
-  const [newCategoryName, setNewCategoryName] = useState(selectedCategoryName);
-
-  const handleNewCategoryNameChange = (e) => {
-    setNewCategoryName(e.target.value);
-  };
-
-  const handleEditCategory = () => {
-    // TODO: 카테고리명을 수정하는 기능을 수행하는 코드 작성
-
-    // 모달 창 닫기
-    console.log(
-      `${selectedCategoryName}(id=${selectedCategoryId})를 ${newCategoryName}으로 update`
-    );
-    onClose();
-  };
-
-  return (
-    <PopupBox>
-      <h2>Edit Category</h2>
-      <input
-        type="text"
-        value={newCategoryName}
-        onChange={handleNewCategoryNameChange}
-      />
-      <button onClick={handleEditCategory}>Save</button>
-      <button onClick={onClose}>Cancel</button>
-    </PopupBox>
-  );
-};
