+++ src/components/AdminPage/adminItems/AdminButtonCSS.jsx
@@ -1,18 +1,38 @@
 import { styled } from "styled-components";
 // import { FaPlus } from "react-icons/fa6";
 
-export const Box = styled.div`
-  height: 100px;
-  width: 100%;
-  background-color: lightblue;
-  margin: 1px;
-`;
 export const PlusBtn = styled.button`
   border: 0;
   height: 50px;
   width: 100%;
   background-color: var(--secondary-color);
 `;
+
+/*
+export const AddButton = styled.button`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 37px;
+  gap: 8px;
+  border-radius: 3px;
+  font-weight: bold;
+  background: var(--secondary-color);
+  cursor: pointer;
+  margin: 6px 17px 25px 17px; // 상 우 하 좌
+
+  .plus {
+    margin-top: 2px;
+    border-radius: 3px;
+    width: 13px;
+    height: 13px;
+    background-color: var(--primary-color);
+    color: white;
+    padding: 2px;
+  }
+`;
+*/
+
 export const Btn = styled.button`
   border: 0;
   border-radius: 10px;
@@ -25,14 +45,4 @@ export const Btn = styled.button`
 export const DeleteBtn = styled(Btn)`
   background-color: #ff4949;
   color: white;
-`;
-
-export const PopupBox = styled.div`
-  background-color: white;
-  border: 1px solid var(--primary-color);
-  width: 40%;
-  height: 20%;
-  position: fixed;
-  left: 0;
-  right: 0;
 `;
